Fix handling invalid length string in Decode()
Currently the code Panic()s when it finds a string
with an invalid length. This isn't ideal because this
functionality can't be tested with a recover() in the
unit tests since it's run in a separate goroutine.

I also can't return an error from a goroutine for obvious reasons,
so I may need to pass an additional channel to the goroutine
just to communicate errors? Not sure if there's a better solution.